cancel and sold out badge in event list
Public event list prompte cancelled and sold out badge for more clarity
to the public users.

Correction for « sold_out » test in home, show and public list. 0
subscriber with 0 free_place should be a free entry case and it appears
to be sold out.

Another error occurs (negative free place)  if an admin change the
number of free places or decide to go free entry with an event.

Home and public list better display with access and price separate.